refactor: move factory files into "factories" folder
