+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './configs';
+export * from './factories/project-config';
+export * from './factories/standard-config';
+export * from './factories/workspace-config';
 export { GLOB_HTML, GLOB_JS, GLOB_SRC, GLOB_TESTS, GLOB_TS } from './globs';
-export * from './project-config';
-export * from './standard-config';
 export type * from './types';
 export * from './utils';
-export * from './workspace-config';
